correcciones en los models y el admin
