--- idgs9A/services/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from .models import Service
-# Register your models here.
-class ServiceAdmin(admin.ModelAdmin):
-    readonly_fields = ('created', 'updated')
-
-admin.site.register(Service, ServiceAdmin)
