--- idgs9A/services/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class ServicesConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'services'
-    verbose_name = 'Gestor de servicios'
