+++ idgs9A/videogames/apps.py
@@ -4,4 +4,5 @@ from django.apps import AppConfig
 class VideogamesConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'videogames'
+    verbose_name = 'Gestor de videojuegos'
